fix bug: delete func + added reload page form

+++ frontend/src/components/Card.jsx
@@ -59,15 +59,21 @@ function Card() {
   }, [editedTask]);
 
   const handleDeleteTask = async () => {
-    console.log(taskData.id);
-    try {
-      await axios.delete(
-        `https://64e48df4c555638029136b4f.mockapi.io/tasks/${taskData.id}`
-      );
-    } catch (error) {
-      console.error("Error deleting task:", error);
+    if (isEditing) {
+      console.log("taskData", taskData.id);
+      const taskId = taskData.id;
+      try {
+        await axios.delete(
+          `https://64e48df4c555638029136b4f.mockapi.io/tasks/${taskId}`
+        );
+      } catch (error) {
+        console.error("Error deleting task:", error);
+      }
     }
   };
+  useEffect(() => {
+    handleDeleteTask();
+  }, [isEditing]);
 
   const handleCancelEditedTask = () => {
     setIsEditing(false);
